+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# 1.3.1
+* Added an `accepts` option to `pushState`.
+* Added optional directory listings.
+
 # 1.3.0-alpha+1
 * ETags once again only encode the first 50 bytes of files. Resolves [#27](https://github.com/angel-dart/static/issues/27).